requirements.txt
openai
argparse
pyyaml
tqdm